Add files via upload
The earlier attack on RSA is successful because of the small exponent and the fact that the same message is
broadcast to three different parties.The same logic has been applied to solve ‘Alice’s Birthday Party (Part2). Hex value of the message was obtained and then converted to ASCII and then Base64 decoded to obtain the output as: Einladung zu
meiner Geburtstagsparty. Die Party findet am 20.12.2012 in Bletchley statt.